Split tests for memory and local storage, increase coverage
